Stevens's Part; main.cpp
The program almost does everything it's required to do, but there's an issue with running the program with MobyDick.txt, also if you parse another file, it'll print 'enter the file's name" twice before letting you enter the file name; entering the file's name works, though. Then, Ally, I've made a count of word and letter percentages in order to run a format.